Update bash script to run Knapsack Pro
We made a few experiments with forked repositeries after applying commit
dc1e0209a and everything seems to work fine, but it looks like that
isn't the case. For instance, on pull requests opened by external
contributors, the tests aren't running.

The fix we're applying is mentioned on the Knapsack Pro repository [1]
(pull request 197).

[1] https://github.com/KnapsackPro/knapsack_pro-ruby