Use Debian Bullseye instead of Buster on Docker
Debian Bullseye was released in 2021 and Debian Buster only receives
security updates.

Note that the libindicator7 package is no longer available in this
distribution. However, we were only installing it in order to install
chromedriver, which we don't do since commit 59b625a5f, so we don't need
it anymore.